lee / temp 12:17

+++ src/main/java/org/iclass/PCProject/product/dto/ReviewDTO.java
@@ -19,7 +19,7 @@ public class ReviewDTO {
     private String username;
     private String code;
     private String content;
-    private long score;
+    private int score;
     private int pSeq;
     private LocalDateTime regDate;
